# backend/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LOUIE AI API",
    "description": "45+ AI-powered endpoints for text processing, code analysis, content generation, and business intelligence. Built by SimplyLouie.",
    "version": "1.0.0",
    "contact": {
      "email": "support@simplylouie.com"
    }
  },
  "servers": [
    {
      "url": "https://louie-api.p.rapidapi.com",
      "description": "RapidAPI Production"
    }
  ],
...
  "security": [
    { "RapidAPIKey": [] }
  ],
  "paths": {
...
    "/v1/pricing": {
      "get": {
        "tags": ["Data"],
        "summary": "Get pricing tiers",
        "description": "Returns the available pricing tiers and their features. No authentication required.",
        "operationId": "getPricing",
        "security": [],
        "responses": {
          "200": {
            "description": "Pricing tiers retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponseObject" },
                "example": {
                  "success": true,
                  "message": "Pricing tiers retrieved successfully",
                  "data": {
                    "result": {
                      "tiers": [
                        { "name": "Free", "price": 0, "requests_per_month": 100, "features": ["Text endpoints", "Basic code endpoints"] },
                        { "name": "Pro", "price": 29, "requests_per_month": 5000, "features": ["All Free features", "Income prediction", "Fraud scan", "Content generation"] },
                        { "name": "Enterprise", "price": 99, "requests_per_month": 50000, "features": ["All Pro features", "Document analysis", "Priority support", "Custom integrations"] }
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/v1/employer/lookup": {
      "get": {
        "tags": ["Data"],
        "summary": "Employer lookup",
        "description": "Looks up employer information by name. Requires authentication.",
        "operationId": "employerLookup",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "description": "The employer name to search for",
            "schema": {
              "type": "string",
              "example": "Google"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employer found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponseObject" },
                "example": {
                  "success": true,
                  "message": "Employer found",
                  "data": {
                    "result": {
                      "name": "Google LLC",
                      "industry": "Technology",
                      "size": "10000+",
                      "headquarters": "Mountain View, CA",
                      "founded": 1998
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "Employer not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "success": false,
                  "message": "Employer not found",
                  "data": null
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/v1/region/data": {
      "get": {
        "tags": ["Data"],
        "summary": "Regional data",
        "description": "Returns regional data for the specified country. Requires authentication.",
        "operationId": "getRegionData",
        "parameters": [
          {
            "name": "country",
            "in": "query",
            "required": true,
            "description": "The country code or name to retrieve data for",
            "schema": {
              "type": "string",
              "example": "US"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Regional data retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponseObject" },
                "example": {
                  "success": true,
                  "message": "Regional data retrieved successfully",
                  "data": {
                    "result": {
                      "country": "United States",
                      "code": "US",
                      "regions": 50,
                      "population": 331000000,
                      "currency": "USD"
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/v1/income/predict": {
      "post": {
        "tags": ["Data"],
        "summary": "Income prediction",
        "description": "Predicts income based on the provided data. Requires Pro+ tier and authentication.",
        "operationId": "predictIncome",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/IncomePredictRequest" },
              "example": {
                "job_title": "Software Engineer",
                "location": "San Francisco, CA",
                "experience_years": 5,
                "education": "Bachelor's"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Income predicted successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponseObject" },
                "example": {
                  "success": true,
                  "message": "Income predicted successfully",
                  "data": {
                    "result": {
                      "predicted_income": 165000,
                      "range": { "low": 140000, "high": 195000 },
                      "confidence": 0.85,
                      "currency": "USD"
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/v1/intelligence/scan": {
      "post": {
        "tags": ["Data"],
        "summary": "Fraud scan",
        "description": "Scans the provided data for potential fraud indicators. Requires Pro+ tier and authentication.",
        "operationId": "fraudScan",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FraudScanRequest" },
              "example": {
                "entity": "John Doe",
                "email": "john.doe@example.com",
                "phone": "+1-555-0123",
                "context": "New account registration"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Fraud scan completed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponseObject" },
                "example": {
                  "success": true,
                  "message": "Fraud scan completed",
                  "data": {
                    "result": {
                      "risk_score": 0.15,
                      "risk_level": "low",
                      "flags": [],
                      "recommendation": "Proceed with standard verification."
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/v1/docs/analyze": {
      "post": {
        "tags": ["Data"],
        "summary": "Document analysis",
        "description": "Analyzes an uploaded document and extracts key information. Requires Enterprise tier and authentication.",
        "operationId": "analyzeDocument",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DocAnalyzeRequest" },
              "example": {
                "document_url": "https://example.com/document.pdf",
                "analysis_type": "summary",
                "language": "en"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document analyzed successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponseObject" },
                "example": {
                  "success": true,
                  "message": "Document analyzed successfully",
                  "data": {
                    "result": {
                      "summary": "The document outlines a partnership agreement between two parties...",
                      "key_entities": ["Company A", "Company B"],
                      "page_count": 12,
                      "language": "en",
                      "topics": ["legal", "partnership", "agreement"]
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "RapidAPIKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RapidAPI-Key",
        "description": "RapidAPI key for authentication. Obtain your key from https://rapidapi.com"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request - missing or invalid parameters",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "success": false,
              "message": "Missing required field: text",
              "data": null
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "success": false,
              "message": "An unexpected error occurred. Please try again later.",
              "data": null
            }
          }
        }
      }
    },
    "schemas": {
...
      "SuccessResponseObject": {
        "type": "object",
        "required": ["success", "message", "data"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Request completed successfully"
          },
          "data": {
            "type": "object",
            "required": ["result"],
            "properties": {
              "result": {
                "type": "object",
                "description": "The result of the AI operation as a structured object"
              }
            }
          }
        }
      },
...
    }
  }
}