# backend/routes/publicApi.js
/**
 * PUBLIC API ROUTES
 * Monetized API endpoints with API key authentication
 * Tiered pricing: Free (100 calls/day), Pro (10K calls/day), Enterprise (unlimited)
 */

const express = require('express');
const { v4: uuidv4 } = require('uuid');
const crypto = require('crypto');

function createPublicApiRouter(db, options = {}) {
  const router = express.Router();

  // ============================================
  // DATABASE TABLES
  // ============================================

  db.exec(`
    -- API Keys table
    CREATE TABLE IF NOT EXISTS api_keys (
      id TEXT PRIMARY KEY,
      key_hash TEXT UNIQUE NOT NULL,
      key_prefix TEXT NOT NULL,
      name TEXT NOT NULL,
      owner_email TEXT NOT NULL,
      tier TEXT DEFAULT 'free',
      rate_limit INTEGER DEFAULT 100,
      is_active INTEGER DEFAULT 1,
      created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
      last_used_at DATETIME
    );

    -- API Usage table
    CREATE TABLE IF NOT EXISTS api_usage (
      id TEXT PRIMARY KEY,
      api_key_id TEXT NOT NULL,
      endpoint TEXT NOT NULL,
      method TEXT NOT NULL,
      response_code INTEGER,
      response_time INTEGER,
      ip_address TEXT,
      created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
      FOREIGN KEY (api_key_id) REFERENCES api_keys(id) ON DELETE CASCADE
    );

    -- API Usage daily aggregate
    CREATE TABLE IF NOT EXISTS api_usage_daily (
      id TEXT PRIMARY KEY,
      api_key_id TEXT NOT NULL,
      date TEXT NOT NULL,
      call_count INTEGER DEFAULT 0,
      FOREIGN KEY (api_key_id) REFERENCES api_keys(id) ON DELETE CASCADE,
      UNIQUE(api_key_id, date)
    );

    -- Create indexes
    CREATE INDEX IF NOT EXISTS idx_api_usage_key ON api_usage(api_key_id);
    CREATE INDEX IF NOT EXISTS idx_api_usage_created ON api_usage(created_at);
    CREATE INDEX IF NOT EXISTS idx_api_daily_key_date ON api_usage_daily(api_key_id, date);
  `);

  // ============================================
  // PREPARED STATEMENTS
  // ============================================

  const statements = {
    getKeyByHash: db.prepare('SELECT * FROM api_keys WHERE key_hash = ?'),
    getKeyById: db.prepare('SELECT * FROM api_keys WHERE id = ?'),
    getKeysByEmail: db.prepare('SELECT id, key_prefix, name, tier, rate_limit, is_active, created_at, last_used_at FROM api_keys WHERE owner_email = ?'),
    createKey: db.prepare('INSERT INTO api_keys (id, key_hash, key_prefix, name, owner_email, tier, rate_limit) VALUES (?, ?, ?, ?, ?, ?, ?)'),
    updateKeyLastUsed: db.prepare('UPDATE api_keys SET last_used_at = CURRENT_TIMESTAMP WHERE id = ?'),
    deactivateKey: db.prepare('UPDATE api_keys SET is_active = 0 WHERE id = ?'),
    activateKey: db.prepare('UPDATE api_keys SET is_active = 1 WHERE id = ?'),
    deleteKey: db.prepare('DELETE FROM api_keys WHERE id = ?'),
    updateKeyTier: db.prepare('UPDATE api_keys SET tier = ?, rate_limit = ? WHERE id = ?'),

    logUsage: db.prepare('INSERT INTO api_usage (id, api_key_id, endpoint, method, response_code, response_time, ip_address) VALUES (?, ?, ?, ?, ?, ?, ?)'),
    getDailyUsage: db.prepare('SELECT * FROM api_usage_daily WHERE api_key_id = ? AND date = ?'),
    upsertDailyUsage: db.prepare('INSERT INTO api_usage_daily (id, api_key_id, date, call_count) VALUES (?, ?, ?, 1) ON CONFLICT(api_key_id, date) DO UPDATE SET call_count = call_count + 1'),

    getUsageStats: db.prepare(`
      SELECT
        COUNT(*) as total_calls,
        AVG(response_time) as avg_response_time,
        SUM(CASE WHEN response_code >= 400 THEN 1 ELSE 0 END) as error_count
      FROM api_usage
      WHERE api_key_id = ? AND created_at >= ?
    `)
  };

  // ============================================
  // HELPER FUNCTIONS
  // ============================================

  function apiResponse(success, message, data = null) {
    return { success, message, data };
  }

  function generateApiKey() {
    const prefix = 'sl_live_';
    const randomPart = crypto.randomBytes(24).toString('hex');
    return prefix + randomPart;
  }

  function hashApiKey(key) {
    return crypto.createHash('sha256').update(key).digest('hex');
  }

  const TIER_LIMITS = {
    free: 100,
    starter: 1000,
    pro: 10000,
    enterprise: 1000000
  };

  const TIER_PRICING = {
    free: { price: 0, name: 'Free', calls: 100 },
    starter: { price: 19, name: 'Starter', calls: 1000 },
    pro: { price: 99, name: 'Pro', calls: 10000 },
    enterprise: { price: 499, name: 'Enterprise', calls: 1000000 }
  };

  // ============================================
  // API KEY AUTH MIDDLEWARE
  // ============================================

  function apiKeyAuth(req, res, next) {
    const apiKey = req.headers['x-api-key'] || req.query.api_key;

    if (!apiKey) {
      return res.status(401).json(apiResponse(false, 'API key required. Get one at simplylouie.com/api'));
    }

    const keyHash = hashApiKey(apiKey);
    const keyRecord = statements.getKeyByHash.get(keyHash);

    if (!keyRecord) {
      return res.status(401).json(apiResponse(false, 'Invalid API key'));
    }

    if (!keyRecord.is_active) {
      return res.status(403).json(apiResponse(false, 'API key has been deactivated'));
    }

    // Check rate limit
    const today = new Date().toISOString().split('T')[0];
    const dailyUsage = statements.getDailyUsage.get(keyRecord.id, today);
    const currentUsage = dailyUsage ? dailyUsage.call_count : 0;

    if (currentUsage >= keyRecord.rate_limit) {
      return res.status(429).json(apiResponse(false, 'Rate limit exceeded. Upgrade your plan at simplylouie.com/api', {
        current_usage: currentUsage,
        limit: keyRecord.rate_limit,
        tier: keyRecord.tier,
        upgrade_url: 'https://simplylouie.com/api/upgrade'
      }));
    }

    // Update usage
    statements.upsertDailyUsage.run(uuidv4(), keyRecord.id, today);
    statements.updateKeyLastUsed.run(keyRecord.id);

    // Attach to request
    req.apiKey = keyRecord;
    req.usageCount = currentUsage + 1;

    // Log usage after response
    const startTime = Date.now();
    res.on('finish', () => {
      const responseTime = Date.now() - startTime;
      const ip = req.ip || req.connection.remoteAddress;
      statements.logUsage.run(uuidv4(), keyRecord.id, req.path, req.method, res.statusCode, responseTime, ip);
    });

    next();
  }

  // Use unified auth middleware if provided (supports RapidAPI + direct keys),
  // otherwise fall back to the local apiKeyAuth above.
  const authMiddleware = options.unifiedAuth || apiKeyAuth;

  // ============================================
  // PUBLIC API ENDPOINTS (No Auth - Info)
  // ============================================

  /**
   * GET /v1
   * API documentation endpoint
   */
  router.get('/', (req, res) => {
    res.json(apiResponse(true, 'SimplyLouie Public API v1', {
      version: '1.0.0',
      documentation: 'https://simplylouie.com/api/docs',
      endpoints: {
        'GET /v1/intelligence/scan': 'Analyze documents for fraud signals (Pro+)',
        'GET /v1/income/predict': 'Predict income from limited data (Pro+)',
        'GET /v1/employer/lookup': 'Employer verification (Free)',
        'GET /v1/region/data': 'Regional economic data (Free)',
        'POST /v1/docs/analyze': 'Full document analysis (Enterprise)'
      },
      pricing: TIER_PRICING,
      get_api_key: 'https://simplylouie.com/api/keys'
    }));
  });

  /**
   * GET /v1/pricing
   * Get pricing information
   */
  router.get('/pricing', (req, res) => {
    res.json(apiResponse(true, 'API Pricing', {
      plans: Object.entries(TIER_PRICING).map(([key, val]) => ({
        tier: key,
        name: val.name,
        price_monthly: val.price,
        calls_per_day: val.calls,
        price_per_call: val.price > 0 ? (val.price / (val.calls * 30)).toFixed(4) : 'free'
      })),
      features_by_tier: {
        free: ['Employer lookup', 'Regional data', 'Basic analytics'],
        starter: ['Everything in Free', 'Income predictions', 'Extended rate limits'],
        pro: ['Everything in Starter', 'Document scanning', 'Fraud detection', 'Priority support'],
        enterprise: ['Everything in Pro', 'Unlimited calls', 'Custom integrations', 'SLA', 'Dedicated support']
      }
    }));
  });

  // ============================================
  // API KEY MANAGEMENT (Requires email verification)
  // ============================================

  /**
   * POST /v1/keys/create
   * Create a new API key
   */
  router.post('/keys/create', (req, res) => {
    try {
      const { email, name, tier = 'free' } = req.body;

      if (!email || !name) {
        return res.status(400).json(apiResponse(false, 'Email and name are required'));
      }

      // Validate email format
      const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
      if (!emailRegex.test(email)) {
        return res.status(400).json(apiResponse(false, 'Invalid email format'));
      }

      // Check existing keys for this email (limit to 5 free keys)
      const existingKeys = statements.getKeysByEmail.all(email);
      const freeKeys = existingKeys.filter(k => k.tier === 'free');
      if (tier === 'free' && freeKeys.length >= 5) {
        return res.status(400).json(apiResponse(false, 'Maximum 5 free API keys per email. Upgrade to create more.'));
      }

      // Generate key
      const apiKey = generateApiKey();
      const keyHash = hashApiKey(apiKey);
      const keyPrefix = apiKey.substring(0, 12) + '...';
      const keyId = uuidv4();
      const rateLimit = TIER_LIMITS[tier] || TIER_LIMITS.free;

      statements.createKey.run(keyId, keyHash, keyPrefix, name, email, tier, rateLimit);

      res.status(201).json(apiResponse(true, 'API key created successfully', {
        key: apiKey,  // Only shown once!
        key_id: keyId,
        name,
        tier,
        rate_limit: rateLimit,
        important: 'Save this key now! It will not be shown again.'
      }));
    } catch (error) {
      console.error('Create key error:', error);
      res.status(500).json(apiResponse(false, 'Failed to create API key'));
    }
  });

  /**
   * GET /v1/keys
   * List API keys for an email (masked)
   */
  router.get('/keys', (req, res) => {
    try {
      const { email } = req.query;

      if (!email) {
        return res.status(400).json(apiResponse(false, 'Email parameter required'));
      }

      const keys = statements.getKeysByEmail.all(email);

      res.json(apiResponse(true, 'API keys retrieved', {
        keys: keys.map(k => ({
          id: k.id,
          prefix: k.key_prefix,
          name: k.name,
          tier: k.tier,
          rate_limit: k.rate_limit,
          is_active: k.is_active === 1,
          created_at: k.created_at,
          last_used_at: k.last_used_at
        })),
        total: keys.length
      }));
    } catch (error) {
      console.error('List keys error:', error);
      res.status(500).json(apiResponse(false, 'Failed to list API keys'));
    }
  });

  /**
   * DELETE /v1/keys/:id
   * Deactivate an API key
   */
  router.delete('/keys/:id', (req, res) => {
    try {
      const { email } = req.body;
      const keyId = req.params.id;

      if (!email) {
        return res.status(400).json(apiResponse(false, 'Email required for verification'));
      }

      const key = statements.getKeyById.get(keyId);
      if (!key || key.owner_email !== email) {
        return res.status(404).json(apiResponse(false, 'API key not found'));
      }

      statements.deactivateKey.run(keyId);

      res.json(apiResponse(true, 'API key deactivated'));
    } catch (error) {
      console.error('Delete key error:', error);
      res.status(500).json(apiResponse(false, 'Failed to deactivate API key'));
    }
  });

  // ============================================
  // USAGE & ANALYTICS
  // ============================================

  /**
   * GET /v1/usage
   * Get usage statistics for an API key
   */
  router.get('/usage', authMiddleware, (req, res) => {
    try {
      const keyId = req.apiKey.id;
      const today = new Date().toISOString().split('T')[0];

      // Today's usage
      const todayUsage = statements.getDailyUsage.get(keyId, today);

      // Last 7 days stats
      const weekAgo = new Date(Date.now() - 7 * 24 * 60 * 60 * 1000).toISOString();
      const weekStats = statements.getUsageStats.get(keyId, weekAgo);

      res.json(apiResponse(true, 'Usage statistics', {
        key: {
          id: req.apiKey.id,
          name: req.apiKey.name,
          tier: req.apiKey.tier,
          rate_limit: req.apiKey.rate_limit
        },
        today: {
          calls: todayUsage ? todayUsage.call_count : 0,
          remaining: req.apiKey.rate_limit - (todayUsage ? todayUsage.call_count : 0),
          limit: req.apiKey.rate_limit
        },
        last_7_days: {
          total_calls: weekStats ? weekStats.total_calls : 0,
          avg_response_time_ms: weekStats ? Math.round(weekStats.avg_response_time) : 0,
          error_count: weekStats ? weekStats.error_count : 0,
          error_rate: weekStats && weekStats.total_calls > 0
            ? ((weekStats.error_count / weekStats.total_calls) * 100).toFixed(2) + '%'
            : '0%'
        }
      }));
    } catch (error) {
      console.error('Usage stats error:', error);
      res.status(500).json(apiResponse(false, 'Failed to get usage statistics'));
    }
  });

  // ============================================
  // INTELLIGENCE API ENDPOINTS (Paid Features)
  // ============================================

  /**
   * GET /v1/employer/lookup
   * Free: Employer verification
   */
  router.get('/employer/lookup', authMiddleware, (req, res) => {
    try {
      const { name, ein } = req.query;

      if (!name && !ein) {
        return res.status(400).json(apiResponse(false, 'Employer name or EIN required'));
      }

      // Simulated employer lookup
      const employers = {
        'google': { name: 'Google LLC', verified: true, industry: 'Technology', employees: '150000+', rating: 'A+' },
        'amazon': { name: 'Amazon.com Inc', verified: true, industry: 'E-commerce/Cloud', employees: '1500000+', rating: 'A+' },
        'walmart': { name: 'Walmart Inc', verified: true, industry: 'Retail', employees: '2300000+', rating: 'A' },
        'default': { name: name || 'Unknown', verified: false, industry: 'Unknown', employees: 'Unknown', rating: 'Not rated' }
      };

      const searchKey = (name || '').toLowerCase();
      const result = employers[searchKey] || employers.default;

      res.json(apiResponse(true, 'Employer lookup complete', {
        employer: result,
        lookup_type: ein ? 'ein' : 'name',
        confidence: result.verified ? 0.95 : 0.3,
        api_calls_remaining: req.apiKey.rate_limit - req.usageCount
      }));
    } catch (error) {
      console.error('Employer lookup error:', error);
      res.status(500).json(apiResponse(false, 'Employer lookup failed'));
    }
  });

  /**
   * GET /v1/region/data
   * Free: Regional economic data
   */
  router.get('/region/data', authMiddleware, (req, res) => {
    try {
      const { country, state } = req.query;

      if (!country) {
        return res.status(400).json(apiResponse(false, 'Country code required'));
      }

      // Simulated regional data
      const regionData = {
        US: {
          country: 'United States',
          currency: 'USD',
          median_income: 65000,
          cost_of_living_index: 100,
          unemployment_rate: 3.7,
          min_wage: 7.25,
          tax_rate_range: '10%-37%'
        },
        UK: {
          country: 'United Kingdom',
          currency: 'GBP',
          median_income: 31000,
          cost_of_living_index: 95,
          unemployment_rate: 4.2,
          min_wage: 10.42,
          tax_rate_range: '20%-45%'
        },
        NG: {
          country: 'Nigeria',
          currency: 'NGN',
          median_income: 2400000,
          cost_of_living_index: 35,
          unemployment_rate: 33.3,
          min_wage: 30000,
          tax_rate_range: '7%-24%'
        }
      };

      const data = regionData[country.toUpperCase()] || {
        country: country.toUpperCase(),
        currency: 'Unknown',
        median_income: null,
        cost_of_living_index: null,
        unemployment_rate: null,
        min_wage: null,
        tax_rate_range: 'Unknown'
      };

      res.json(apiResponse(true, 'Regional data retrieved', {
        region: data,
        data_source: 'SimplyLouie Intelligence',
        last_updated: new Date().toISOString().split('T')[0],
        api_calls_remaining: req.apiKey.rate_limit - req.usageCount
      }));
    } catch (error) {
      console.error('Region data error:', error);
      res.status(500).json(apiResponse(false, 'Region data lookup failed'));
    }
  });

  /**
   * POST /v1/income/predict
   * Pro: Income prediction from limited data
   */
  router.post('/income/predict', authMiddleware, (req, res) => {
    try {
      // Check tier
      if (!['pro', 'enterprise'].includes(req.apiKey.tier)) {
        return res.status(403).json(apiResponse(false, 'Pro tier required for income predictions', {
          current_tier: req.apiKey.tier,
          upgrade_url: 'https://simplylouie.com/api/upgrade'
        }));
      }

      const { job_title, industry, location, experience_years, education } = req.body;

      if (!job_title) {
        return res.status(400).json(apiResponse(false, 'Job title required'));
      }

      // Simulated income prediction
      const baseSalary = 50000;
      const industryMultiplier = { 'technology': 1.5, 'finance': 1.4, 'healthcare': 1.2, 'retail': 0.8, 'default': 1.0 };
      const expMultiplier = Math.min(1 + (experience_years || 0) * 0.05, 2.0);
      const eduMultiplier = { 'phd': 1.3, 'masters': 1.2, 'bachelors': 1.1, 'default': 1.0 };

      const indMult = industryMultiplier[industry?.toLowerCase()] || industryMultiplier.default;
      const eduMult = eduMultiplier[education?.toLowerCase()] || eduMultiplier.default;

      const predictedSalary = Math.round(baseSalary * indMult * expMultiplier * eduMult);

      res.json(apiResponse(true, 'Income prediction complete', {
        prediction: {
          annual_salary_low: Math.round(predictedSalary * 0.8),
          annual_salary_mid: predictedSalary,
          annual_salary_high: Math.round(predictedSalary * 1.25),
          monthly_gross: Math.round(predictedSalary / 12),
          monthly_net_estimate: Math.round((predictedSalary / 12) * 0.75),
          confidence: 0.72
        },
        factors: {
          job_title,
          industry: industry || 'general',
          experience_years: experience_years || 0,
          education: education || 'not specified',
          location: location || 'US average'
        },
        api_calls_remaining: req.apiKey.rate_limit - req.usageCount
      }));
    } catch (error) {
      console.error('Income predict error:', error);
      res.status(500).json(apiResponse(false, 'Income prediction failed'));
    }
  });

  /**
   * POST /v1/docs/analyze
   * Enterprise: Full document analysis
   */
  router.post('/docs/analyze', authMiddleware, (req, res) => {
    try {
      // Check tier
      if (req.apiKey.tier !== 'enterprise') {
        return res.status(403).json(apiResponse(false, 'Enterprise tier required for document analysis', {
          current_tier: req.apiKey.tier,
          upgrade_url: 'https://simplylouie.com/api/upgrade'
        }));
      }

      const { document_type, document_data, options } = req.body;

      if (!document_type || !document_data) {
        return res.status(400).json(apiResponse(false, 'document_type and document_data required'));
      }

      // Simulated document analysis
      const analysisId = uuidv4();
      const riskScore = Math.random() * 100;

      const fraudSignals = [];
      if (riskScore > 70) fraudSignals.push({ type: 'inconsistent_metadata', severity: 'medium' });
      if (riskScore > 85) fraudSignals.push({ type: 'potential_manipulation', severity: 'high' });

      res.json(apiResponse(true, 'Document analysis complete', {
        analysis_id: analysisId,
        document_type,
        results: {
          risk_score: Math.round(riskScore * 100) / 100,
          risk_level: riskScore < 30 ? 'low' : riskScore < 70 ? 'medium' : 'high',
          fraud_signals: fraudSignals,
          extracted_data: {
            document_date: new Date().toISOString().split('T')[0],
            issuer: 'Detected Issuer',
            amounts: [1234.56, 2345.67],
            names: ['John Doe']
          },
          income_estimate: {
            monthly: 5432,
            annual: 65184,
            stability: riskScore < 50 ? 'stable' : 'variable'
          },
          recommendations: riskScore > 60
            ? ['Request additional verification', 'Cross-reference with employer']
            : ['Document verified', 'Proceed with processing']
        },
        processing_time_ms: Math.floor(Math.random() * 500) + 200,
        api_calls_remaining: req.apiKey.rate_limit - req.usageCount
      }));
    } catch (error) {
      console.error('Doc analyze error:', error);
      res.status(500).json(apiResponse(false, 'Document analysis failed'));
    }
  });

  /**
   * POST /v1/intelligence/scan
   * Pro: Quick fraud scan
   */
  router.post('/intelligence/scan', authMiddleware, (req, res) => {
    try {
      // Check tier
      if (!['pro', 'enterprise'].includes(req.apiKey.tier)) {
        return res.status(403).json(apiResponse(false, 'Pro tier required for intelligence scanning', {
          current_tier: req.apiKey.tier,
          upgrade_url: 'https://simplylouie.com/api/upgrade'
        }));
      }

      const { data_points } = req.body;

      if (!data_points || !Array.isArray(data_points)) {
        return res.status(400).json(apiResponse(false, 'data_points array required'));
      }

      // Analyze data points for fraud signals
      const signals = [];
      let totalRisk = 0;

      data_points.forEach((point, i) => {
        const pointRisk = Math.random() * 50;
        totalRisk += pointRisk;
        if (pointRisk > 30) {
          signals.push({
            index: i,
            type: 'anomaly_detected',
            risk: Math.round(pointRisk),
            detail: `Data point ${i} shows unusual patterns`
          });
        }
      });

      const avgRisk = totalRisk / data_points.length;

      res.json(apiResponse(true, 'Intelligence scan complete', {
        scan_id: uuidv4(),
        data_points_analyzed: data_points.length,
        overall_risk: Math.round(avgRisk),
        risk_level: avgRisk < 20 ? 'low' : avgRisk < 40 ? 'medium' : 'high',
        signals: signals.slice(0, 10),
        recommendation: avgRisk > 40 ? 'Manual review recommended' : 'Data appears legitimate',
        api_calls_remaining: req.apiKey.rate_limit - req.usageCount
      }));
    } catch (error) {
      console.error('Intelligence scan error:', error);
      res.status(500).json(apiResponse(false, 'Intelligence scan failed'));
    }
  });

  return router;
}

module.exports = createPublicApiRouter;